# Documentation/src/content/schemas/swagger.yaml
openapi: 3.0.0
info:
  title: Zello API
  version: 1.0.0
  description: |-
    A comprehensive API for managing projects, lists, and cards similar to Trello. This API allows users to:
    - Create and manage workspaces, projects, lists, and cards
    - Organize work using customizable boards and lists
    - Collaborate with team members through comments and mentions
    - Track work progress with activity logs
    - Manage permissions and roles at different levels
    
    ### Features
    - Full CRUD operations for all resources
    - Rich card features including attachments, checklists, and labels
    - Comprehensive activity tracking
    - Fine-grained permission system
    - Authentication and user management
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Project:
      description: Represents a project with its lists and settings
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
          example: "User Authentication Service"
        description:
          type: string
        type:
          type: string
          enum:
            - scrum
            - kanban
            - basic
        startDate:
          type: string
          format: date-time
          description: Project start date
          example: "2024-01-01T00:00:00Z"
        endDate:
          type: string
          format: date-time
          description: Project target completion date
          example: "2024-06-30T23:59:59Z"
        settings:
          type: object
          properties:
            sprintDuration:
              type: integer
            pointScale:
              type: string
            defaultEstimateUnit:
              type: string
            defaultIssueTypes:
              type: array
              description: Default issue types available in this project
              items:
                type: string
                enum: [ task, bug, feature, improvement, epic ]
        status:
          type: string
          enum:
            - planning    # Project is in planning phase
            - active     # Project is currently being worked on
            - onHold    # Project is temporarily paused
            - completed # Project has been completed
            - archived  # Project is archived
        actualStartDate:
          type: string
          format: date-time
          description: Actual date when project started
        actualEndDate:
          type: string
          format: date-time
          description: Actual date when project was completed
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: Invalid request parameters
        code:
          type: string
          example: INVALID_PARAMETERS
security:
  - bearerAuth: [ ]
paths:
...
  /projects/{projectId}:
    get:
      summary: Get project details
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              examples:
                success:
                  summary: Project details response
                  value: {
                    "id": "123e4567-e89b-12d3-a456-426614174030",
                    "workspaceId": "123e4567-e89b-12d3-a456-426614174010",
                    "name": "Website Redesign",
                    "description": "Complete overhaul of company website with modern design and improved UX",
                    "type": "kanban",
                    "startDate": "2024-01-01T00:00:00Z",
                    "endDate": "2024-06-30T23:59:59Z",
                    "actualStartDate": "2024-01-01T09:00:00Z",
                    "actualEndDate": null,
                    "settings": {
                      "sprintDuration": 14,
                      "pointScale": "fibonacci",
                      "defaultEstimateUnit": "days",
                      "columns": [
                        {
                          "name": "Backlog",
                          "limit": null
                        },
                        {
                          "name": "In Progress",
                          "limit": 5
                        },
                        {
                          "name": "Review",
                          "limit": 3
                        },
                        {
                          "name": "Done",
                          "limit": null
                        }
                      ],
                      "labels": [
                        {
                          "name": "frontend",
                          "color": "#FF0000"
                        },
                        {
                          "name": "backend",
                          "color": "#00FF00"
                        },
                        {
                          "name": "urgent",
                          "color": "#FF4500"
                        }
                      ]
                    },
                    "status": "active",
                    "members": [
                      {
                        "userId": "123e4567-e89b-12d3-a456-426614174001",
                        "role": "admin"
                      },
                      {
                        "userId": "123e4567-e89b-12d3-a456-426614174002",
                        "role": "member"
                      }
                    ],
                    "createdAt": "2024-01-01T12:00:00Z",
                    "updatedAt": "2024-01-15T16:30:00Z"
                  }
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Project not found"
                code: "PROJECT_NOT_FOUND"

    put:
      summary: Update project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: "Updated Project Name"
              description: "Updated project description"
              type: "scrum"
              status: "active"
              settings: {
                sprintDuration: 14,
                pointScale: "fibonacci",
                defaultEstimateUnit: "days"
              }
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                type:
                  type: string
                  enum: [ scrum, kanban, basic ]
                status:
                  type: string
                  enum: [ active, archived, completed ]
                settings:
                  type: object
      responses:
        '200':
          description: Project updated
        '403':
          description: Insufficient permissions
        '404':
          description: Project not found

    delete:
      summary: Delete project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Project deleted
        '403':
          description: Insufficient permissions
        '404':
          description: Project not found
  '/projects/{projectId}/lists':
    post:
      summary: Create a new list
      description: Creates a new list within a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                position:
                  type: integer
            example:
              title: In Progress
              position: 1
      responses:
        '201':
          description: List successfully created
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  '/projects/{projectId}/members':
    post:
      summary: Add member to project
      description: Adds a new member to the project with specified role
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - role
              properties:
                userId:
                  type: string
                  format: uuid
                role:
                  type: string
                  enum:
                    - admin
                    - member
                    - viewer
            example:
              userId: "123e4567-e89b-12d3-a456-426614174000"
              role: "member"
              permissions: [
                "manage_cards",
                "comment",
                "view"
              ]
      responses:
        '201':
          description: Member successfully added to project
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'